유저 관련 기능 추가

+++ src/main/java/com/hhplush/eCommerce/api/user/dto/response/ResponseGetUserDTO.java
@@ -4,9 +4,9 @@ import lombok.Builder;
 
 @Builder
 public record ResponseGetUserDTO(
-    Integer userId,
+    Long userId,
     String userName,
-    Integer point
+    Long point
 ) {
 
 }
